+++ docs/source/autoapi_pymor.py
@@ -63,9 +63,9 @@ SKIPS_RE = re.compile(r'(?:{})'.format('|'.join(map(re.escape, sorted(MANUAL_SKI
 def skip(app, what, name, obj, skip, options):
     if name == 'pymor.operators.interface.Operator._assemble_lincomb':
         return False
-    if name == 'pymor.algorithms.lradi.LradiLyapunovSolverLRCF._solve':
+    if name == 'pymor.solvers.matrix.lradi.LradiLyapunovSolverLRCF._solve':
         return False
-    if name == 'pymor.algorithms.lrradi.LrradiRiccatiSolverLRCF._solve':
+    if name == 'pymor.solvers.matrix.lrradi.LrradiRiccatiSolverLRCF._solve':
         return False
     try:
         if ':noindex:' in obj.docstring:
