+++ src/pymordemos/string_equation.py
@@ -38,7 +38,7 @@ def main(n: int = 101, r: int = 5):
     """
     set_log_levels({
         'pymor.algorithms.gram_schmidt.gram_schmidt': 'WARNING',
-        'pymor.algorithms.lradi.solve_lyap_lrcf': 'WARNING',
+        'pymor.solvers.matrix.lradi.solve_lyap_lrcf': 'WARNING',
         'pymor.reductors.basic.LTIPGReductor': 'WARNING',
         'pymor.reductors.basic.SOLTIPGReductor': 'WARNING',
     })
